requirements.txt
python-telegram-bot[job-queue]==22.6
APScheduler==3.11.2
